+++ tools/bugserver-trigger/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @internal/bugserver-trigger
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @visactor/vquery@0.1.48
+  - @visactor/vseed@0.1.48
+
 ## 0.0.2
 
 ### Patch Changes